Related work items: #1875

+++ VerticalWarehouses/Common/Ferretto.VW.CommonUtils/Messages/Interfaces/IMovementMessageData.cs
@@ -8,7 +8,7 @@ namespace Ferretto.VW.CommonUtils.Messages.Interfaces
 
         Axis Axis { get; set; }
 
-        decimal? Displacement { get; set; }
+        double? Displacement { get; set; }
 
         MovementType MovementType { get; set; }
